examples/python_live_project/pyproject.toml
[project]
name = "ecommerce-test-automation"
version = "1.0.0"
description = "E2E Test Automation Framework for EcommerceApp"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "pytest>=7.4.0",
    "pytest-playwright>=0.4.3",
    "pytest-html>=4.1.0",
    "allure-pytest>=2.13.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "faker>=20.0.0",
    "openpyxl>=3.1.0",
    "jsonschema>=4.20.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: Smoke tests",
    "regression: Regression tests",
    "api: API tests",
    "e2e: End to end tests",
    "flaky: Flaky tests that may need retries",
]
addopts = "--strict-markers -v"
